Rename id2vec cmd files to group them together
